requirements.txt
asgiref==3.3.4
Django==3.2.4
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
asgiref==3.3.4
Django==3.2.4
django-cors-headers==3.7.0
django-crispy-forms==1.12.0
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
